change mu and c

+++ src/interfaces/IPool.sol
@@ -19,7 +19,7 @@ interface IPool is IERC20, IERC2612 {
     function fyToken() external view returns(IMaturingToken);
     function g1() external view returns(int128);
     function g2() external view returns(int128);
-    function getC() external view returns (int128);
+    function getC() external view returns (uint256);
     function getCurrentSharePrice() external view returns (uint256);
     function getCache() external view returns (uint104 baseCached, uint104 fyTokenCached, uint32 blockTimestampLast, uint16 g1Fee_);
     function getBaseBalance() external view returns(uint128);
@@ -28,7 +28,7 @@ interface IPool is IERC20, IERC2612 {
     function init(address to) external returns (uint256, uint256, uint256);
     function maturity() external view returns(uint32);
     function mint(address to, address remainder, uint256 minRatio, uint256 maxRatio) external returns (uint256, uint256, uint256);
-    function mu() external view returns (int128);
+    function mu() external view returns (uint256);
     function mintWithBase(address to, address remainder, uint256 fyTokenToBuy, uint256 minRatio, uint256 maxRatio) external returns (uint256, uint256, uint256);
     function retrieveBase(address to) external returns(uint128 retrieved);
     function retrieveFYToken(address to) external returns(uint128 retrieved);
